removed unused init.d script, deployment script works. hurruh!

+++ deployment/deploy.sh
@@ -1,27 +1,33 @@
 
+if [ ! -n "$1" ]
+then
+	echo "Usage: `basename $0` server_name"
+	exit 1
+fi
+
+SERVER_NAME=$1
 
 # remove the old file
 rm RebeccaCheri.zip --force
 
 # create a new deployment archive
-zip -r ./RebeccaCheri.zip ../../Bokkas-Hair-Dressing -x "**/.git/*" -x "**/tmp/*" -x "**/deployment/*" -x "**/eclipse/*" -x "**/vmsetup/*" -x "**/.settings/*" -x "**/test/*" -x "**/.classpath" -x "**/.project" -x "**/.gitignore"
+cd ../..
+zip -r ./Bokkas-Hair-Dressing/deployment/RebeccaCheri.zip Bokkas-Hair-Dressing -x "**/.git/*" -x "**/tmp/*" -x "**/deployment/*" -x "**/eclipse/*" -x "**/vmsetup/*" -x "**/.settings/*" -x "**/test/*" -x "**/.classpath" -x "**/.project" -x "**/.gitignore"
+cd ./Bokkas-Hair-Dressing/deployment
 
 # copy latest artifact to the remote server, unzip
-scp RebeccaCheri.zip root@174.143.168.239:~
+scp RebeccaCheri.zip bokka@$SERVER_NAME:~
 
 # stop the play server
-
+ssh bokka@$SERVER_NAME '/var/lib/play-1.2.1/play stop RebeccaCheri'
 
 # remove the old rebecca cheri
-ssh root@174.143.168.239 'rm RebeccaCheri_previous -rf'
-ssh root@174.143.168.239 'mv RebeccaCheri RebeccaCheri_previous'
+ssh bokka@$SERVER_NAME 'rm RebeccaCheri_previous -rf'
+ssh bokka@$SERVER_NAME 'mv RebeccaCheri RebeccaCheri_previous'
 
 # extract the new rebecca cheri
-ssh root@174.143.168.239 'unzip RebeccaCheri.zip'
-ssh root@174.143.168.239 'mv Bokkas-Hair-Dressing RebeccaCheri'
-
+ssh bokka@$SERVER_NAME 'unzip RebeccaCheri.zip'
+ssh bokka@$SERVER_NAME 'mv Bokkas-Hair-Dressing RebeccaCheri'
 
 # start the play server
-
-
-
+ssh bokka@$SERVER_NAME '/var/lib/play-1.2.1/play start RebeccaCheri'
